refactor: move type definitions to core
Move type definitions from CLI to core, to
make the CLI depends on the core to implement
and not the other way. Type definitions moved
include stack plugins signature and ci platform
writer plugins.
