--- core/plugins/platforms/deps.ts
@@ -1,2 +1,0 @@
-export { ensureFile, log } from "../../../cli/deps.ts";
-export * from "../../../cli/src/platform/plugin.ts";
