+++ core/plugins/stack/_shared/node_package_manager/mod.ts
@@ -1,4 +1,4 @@
-import { IntrospectFn } from "../../deps.ts";
+import { IntrospectFn } from "../../../../types.ts";
 
 interface Npm {
   name: "npm";
