--- core/plugins/stack/deps.ts
@@ -1,1 +1,0 @@
-export * from "../../../cli/src/plugin/mod.ts";
